+++ wallet/api/client_connect_wallet.go
@@ -78,33 +78,59 @@ func (h *ClientConnectWallet) Handle(ctx context.Context, hostname string) (wall
 			return nil, requestInterruptedError(ErrRequestInterrupted)
 		}
 
-		selectedWallet, err := h.interactor.RequestWalletSelection(ctx, traceID, hostname, availableWallets)
-		if err != nil {
-			if errDetails := handleRequestFlowError(ctx, traceID, h.interactor, err); errDetails != nil {
-				return nil, errDetails
+		var walletName string
+		if len(availableWallets) > 1 {
+			selectedWallet, err := h.interactor.RequestWalletSelection(ctx, traceID, hostname, availableWallets)
+			if err != nil {
+				if errDetails := handleRequestFlowError(ctx, traceID, h.interactor, err); errDetails != nil {
+					return nil, errDetails
+				}
+				h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("requesting the wallet selection failed: %w", err))
+				return nil, internalError(ErrCouldNotConnectToWallet)
 			}
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("requesting the wallet selection failed: %w", err))
-			return nil, internalError(ErrCouldNotConnectToWallet)
-		}
 
-		if exist, err := h.walletStore.WalletExists(ctx, selectedWallet.Wallet); err != nil {
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not verify the wallet existence: %w", err))
-			return nil, internalError(ErrCouldNotConnectToWallet)
-		} else if !exist {
-			h.interactor.NotifyError(ctx, traceID, UserError, ErrWalletDoesNotExist)
-			continue
-		}
-
-		if err := h.walletStore.UnlockWallet(ctx, selectedWallet.Wallet, selectedWallet.Passphrase); err != nil {
-			if errors.Is(err, wallet.ErrWrongPassphrase) {
-				h.interactor.NotifyError(ctx, traceID, UserError, wallet.ErrWrongPassphrase)
+			if exist, err := h.walletStore.WalletExists(ctx, selectedWallet); err != nil {
+				h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not verify the wallet existence: %w", err))
+				return nil, internalError(ErrCouldNotConnectToWallet)
+			} else if !exist {
+				h.interactor.NotifyError(ctx, traceID, UserError, ErrWalletDoesNotExist)
 				continue
 			}
-			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not unlock the wallet: %w", err))
+
+			walletName = selectedWallet
+		} else {
+			// There is single wallet available, it doesn't make sense to ask which
+			// wallet to use.
+			walletName = availableWallets[0]
+		}
+
+		alreadyUnlocked, err := h.walletStore.IsWalletAlreadyUnlocked(ctx, walletName)
+		if err != nil {
+			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not verify whether the wallet is already unlock or not: %w", err))
 			return nil, internalError(ErrCouldNotConnectToWallet)
 		}
 
-		w, err := h.walletStore.GetWallet(ctx, selectedWallet.Wallet)
+		if !alreadyUnlocked {
+			walletPassphrase, err := h.interactor.RequestPassphrase(ctx, traceID, walletName)
+			if err != nil {
+				if errDetails := handleRequestFlowError(ctx, traceID, h.interactor, err); errDetails != nil {
+					return nil, errDetails
+				}
+				h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("requesting the wallet passphrase failed: %w", err))
+				return nil, internalError(ErrCouldNotConnectToWallet)
+			}
+
+			if err := h.walletStore.UnlockWallet(ctx, walletName, walletPassphrase); err != nil {
+				if errors.Is(err, wallet.ErrWrongPassphrase) {
+					h.interactor.NotifyError(ctx, traceID, UserError, wallet.ErrWrongPassphrase)
+					continue
+				}
+				h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not unlock the wallet: %w", err))
+				return nil, internalError(ErrCouldNotConnectToWallet)
+			}
+		}
+
+		w, err := h.walletStore.GetWallet(ctx, walletName)
 		if err != nil {
 			h.interactor.NotifyError(ctx, traceID, InternalError, fmt.Errorf("could not retrieve the wallet: %w", err))
 			return nil, internalError(ErrCouldNotConnectToWallet)